default size not large enough to hold gigantic page size
The default page size needs to be stored as a long in order to hold a gigantic page size (e.g. 16GB).

+++ hugeutils.c
@@ -55,7 +55,7 @@ static struct hpage_size hpage_sizes[MAX_HPAGE_SIZES];
 static int nr_hpage_sizes;
 static int hpage_sizes_default_idx = -1;
 
-static int default_size;
+static long default_size;
 
 /********************************************************************/
 /* Internal functions                                               */
